Refactored MenuItems
