+++ src/main/java/rip/shukaaa/Main.java
@@ -10,7 +10,6 @@ import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 
-
 public class Main {
     public static File backupImage = null;
     public static ShukaaaImage img = null;
