+++ src/main/java/rip/shukaaa/ui/components/UiMenuBar.java
@@ -7,16 +7,21 @@ import rip.shukaaa.effect.input.EffectInput;
 import rip.shukaaa.effect.input.inputs.ComboBox;
 import rip.shukaaa.effect.input.inputs.Slider;
 import rip.shukaaa.enums.EffectCategory;
-import rip.shukaaa.enums.ImageFormats;
 import rip.shukaaa.exceptions.EffectOptionNotFoundException;
 import rip.shukaaa.exceptions.ImageNotFoundException;
 import rip.shukaaa.image.Pixel;
 import rip.shukaaa.ui.UiManager;
+import rip.shukaaa.ui.logic.menu.items.edit.RedoMenuItem;
+import rip.shukaaa.ui.logic.menu.items.edit.ResetMenuItem;
+import rip.shukaaa.ui.logic.menu.items.edit.UndoMenuItem;
+import rip.shukaaa.ui.logic.menu.items.file.ExitMenuItem;
+import rip.shukaaa.ui.logic.menu.items.file.OpenMenuItem;
+import rip.shukaaa.ui.logic.menu.items.file.SaveMenuItem;
+import rip.shukaaa.ui.logic.menu.items.image.ResizeMenuItem;
+import rip.shukaaa.utils.UiUtils;
 
 import javax.swing.*;
-import javax.swing.plaf.basic.BasicSpinnerUI;
 import java.awt.*;
-import java.io.File;
 import java.util.*;
 
 public class UiMenuBar extends JMenuBar {
@@ -28,61 +33,13 @@ public class UiMenuBar extends JMenuBar {
         JMenu fileMenu = new JMenu("File");
         this.add(fileMenu);
 
-        JMenuItem open = new JMenuItem("Open");
-        open.addActionListener(e -> {
-            JFileChooser fileChooser = new JFileChooser();
-            fileChooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
-            fileChooser.showOpenDialog(null);
+        JMenuItem open = new OpenMenuItem(uiManager).getItem();
+        JMenuItem save = new SaveMenuItem().getItem();
+        JMenuItem exit = new ExitMenuItem().getItem();
 
-            File selectedFile = fileChooser.getSelectedFile();
-            if (selectedFile == null) {
-                return;
-            }
-
-            try {
-                Main.setImg(selectedFile);
-                Main.backupImage = selectedFile;
-                Main.updateImage(Main.getImg(Main.backupImage), uiManager);
-            } catch (ImageNotFoundException ex) {
-                throw new RuntimeException(ex);
-            }
-        });
         fileMenu.add(open);
-
-        JMenuItem save = new JMenuItem("Save");
-        save.addActionListener(e -> {
-            JDialog saveDialog = createDialog("Save Image");
-            JTextField fileName = new JTextField();
-            fileName.setColumns(10);
-            fileName.setBorder(BorderFactory.createTitledBorder("File Name"));
-            saveDialog.add(fileName);
-            JComboBox<ImageFormats> fileExtension = new JComboBox<>(ImageFormats.values());
-            fileExtension.setBorder(BorderFactory.createEmptyBorder(10, 10, 10, 10));
-            saveDialog.add(fileExtension);
-
-            JSpinner widthSpinner = createSpinner(Main.img.getWidth(), 1, 10000, "Width");
-            saveDialog.add(widthSpinner);
-
-            JSpinner heightSpinner = createSpinner(Main.img.getHeight(), 1, 10000, "Height");
-            saveDialog.add(heightSpinner);
-
-            JButton saveButton = new JButton("Save");
-            saveDialog.add(saveButton);
-            saveButton.addActionListener(e1 -> {
-                Main.img.export(fileName.getText(), (ImageFormats) fileExtension.getSelectedItem(), (int) widthSpinner.getValue(), (int) heightSpinner.getValue());
-                saveDialog.dispose();
-            });
-            saveDialog.pack();
-        });
         fileMenu.add(save);
-
         fileMenu.addSeparator();
-
-        JMenuItem exit = new JMenuItem("Exit");
-        exit.addActionListener(e -> {
-            Main.deleteTempImages();
-            System.exit(0);
-        });
         fileMenu.add(exit);
 
         //*************//
@@ -92,66 +49,40 @@ public class UiMenuBar extends JMenuBar {
         JMenu editMenu = new JMenu("Edit");
         this.add(editMenu);
 
-        JMenuItem reset = new JMenuItem("Reset");
-        reset.addActionListener(e -> {
-            try {
-                Main.updateImage(Main.getImg(Main.backupImage), uiManager);
-            } catch (ImageNotFoundException ex) {
-                throw new RuntimeException(ex);
-            }
-        });
+        JMenuItem reset = new ResetMenuItem(uiManager).getItem();
+        JMenuItem undo = new UndoMenuItem(uiManager).getItem();
+        JMenuItem redo = new RedoMenuItem(uiManager).getItem();
+
         editMenu.add(reset);
-
         editMenu.addSeparator();
-
-        JMenuItem undo = new JMenuItem("Undo");
-        undo.addActionListener(e -> {
-            Main.img.undo();
-            Main.updateImage(Main.img, uiManager);
-        });
         editMenu.add(undo);
-
-        JMenuItem redo = new JMenuItem("Redo");
-        redo.addActionListener(e -> {
-            Main.img.redo();
-            Main.updateImage(Main.img, uiManager);
-        });
         editMenu.add(redo);
 
+        //****************//
+        //* Effects Menu *//
+        //****************//
+
+        // Effects get added dynamically, so we just need to create the menu here
         JMenu effectsMenu = new JMenu("Effects");
         this.add(effectsMenu);
+
+        //***************//
+        //* Image Menu *//
+        //***************//
 
         JMenu imageMenu = new JMenu("Image");
         this.add(imageMenu);
 
-        imageMenu.add(createLabelTitle("Properties: "));
-        JMenuItem resize = new JMenuItem("Resize");
-        resize.addActionListener(e -> {
-            JDialog resizeDialog = createDialog("Resize Image");
-            resizeDialog.add(new JLabel("Width:"));
-            JSpinner widthSpinner = createSpinner(Main.img.getWidth(), 1, 10000, "Width");
-            resizeDialog.add(widthSpinner);
-            resizeDialog.add(new JLabel("Height:"));
-            JSpinner heightSpinner = createSpinner(Main.img.getHeight(), 1, 10000, "Height");
-            resizeDialog.add(heightSpinner);
-            JButton apply = new JButton("Apply");
-            resizeDialog.add(apply);
-            apply.addActionListener(e1 -> {
-                Main.img = Main.img.resizedImage((int) widthSpinner.getValue(), (int) heightSpinner.getValue());
-                Main.updateImage(Main.img, uiManager);
-                resizeDialog.dispose();
-            });
-            resizeDialog.pack();
-        });
+        JLabel propertiesLabel = UiUtils.createLabelTitle("Properties: ");
+        JMenuItem resize = new ResizeMenuItem(uiManager).getItem();
+
+        imageMenu.add(propertiesLabel);
         imageMenu.add(resize);
 
-        // Disable all menu items except for open
-        for (int i = 1; i < this.getMenuCount(); i++) {
-            JMenu menu = this.getMenu(i);
-            menu.setEnabled(false);
-        }
+        //***************//
 
-        save.setEnabled(false);
+        this.disableMenu(save);
+
 
         HashMap<String, Effect> effects = EffectRegister.effects;
         HashMap<EffectCategory, ArrayList<JMenuItem>> items = new HashMap<>();
@@ -174,7 +105,7 @@ public class UiMenuBar extends JMenuBar {
                     return;
                 }
 
-                JDialog dialog = createDialog("Add " + name);
+                JDialog dialog = UiUtils.createDialog("Add " + name);
                 HashMap<String, JComponent> inputMap = new HashMap<>();
 
                 // Create dialog inputs for each effect input
@@ -266,7 +197,7 @@ public class UiMenuBar extends JMenuBar {
                     imageMenu.addSeparator();
                 }
 
-                imageMenu.add(createLabelTitle(category.getSubCategory() + ": "));
+                imageMenu.add(UiUtils.createLabelTitle(category.getSubCategory() + ": "));
                 menuItems.forEach(imageMenu::add);
             }
 
@@ -275,7 +206,7 @@ public class UiMenuBar extends JMenuBar {
                     effectsMenu.addSeparator();
                 }
 
-                effectsMenu.add(createLabelTitle(category.getSubCategory() + ": "));
+                effectsMenu.add(UiUtils.createLabelTitle(category.getSubCategory() + ": "));
                 menuItems.forEach(effectsMenu::add);
             }
         }
@@ -294,35 +225,11 @@ public class UiMenuBar extends JMenuBar {
         }
     }
 
-    private JLabel createLabelTitle(String title) {
-        JLabel label = new JLabel(title);
-        // add bold
-        label.setFont(label.getFont().deriveFont(14.0f).deriveFont(Font.ITALIC));
-        label.setBorder(BorderFactory.createEmptyBorder(0, 8, 4, 0));
-        return label;
-    }
-
-    private JDialog createDialog(String title) {
-        JDialog dialog = new JDialog();
-        dialog.setLayout(new BoxLayout(dialog.getContentPane(), BoxLayout.Y_AXIS));
-        dialog.setTitle(title);
-        dialog.setVisible(true);
-        return dialog;
-    }
-
-    private JSpinner createSpinner(int value, int min, int max, String title) {
-        SpinnerModel spinnerModel = new SpinnerNumberModel(value, min, max, 1);
-        JSpinner spinner = new JSpinner(spinnerModel);
-        spinner.setBorder(BorderFactory.createTitledBorder(title));
-        spinner.setUI(new BasicSpinnerUI() {
-            protected Component createNextButton() {
-                return null;
-            }
-
-            protected Component createPreviousButton() {
-                return null;
-            }
-        });
-        return spinner;
+    private void disableMenu(JMenuItem save) {
+        for (int i = 1; i < this.getMenuCount(); i++) {
+            JMenu menu = this.getMenu(i);
+            menu.setEnabled(false);
+        }
+        save.setEnabled(false);
     }
 }
