+++ src/main/java/rip/shukaaa/ui/logic/menu/items/MenuItem.java
@@ -1,0 +1,17 @@
+package rip.shukaaa.ui.logic.menu.items;
+
+import javax.swing.*;
+
+public abstract class MenuItem {
+		JMenuItem item;
+
+		protected abstract JMenuItem createItem();
+
+		public JMenuItem getItem() {
+				if (item == null) {
+						item = createItem();
+				}
+
+				return item;
+		}
+}
